LAD-136 #time 1h 30m #comment Implemented starting a new driving session when entering drive fragment, if connected to mower. On back button will save the current driving session's events

+++ LaderAppen2/app/src/main/java/se/ju/student/hihe1788/laderappen2/RouteModel.kt
@@ -15,6 +15,6 @@ class RouteModel(val id: Int, val positionEvents: ArrayList<PointModel> = ArrayL
 }
 
 
-class PointModel(val eventId: Int? = null, val eventTypeId: Int, val positionX: Int, val positionY: Int, val dateCreated: Date) {
+class PointModel(val eventId: Int? = null, val eventTypeId: Int, val positionX: Float, val positionY: Float, val dateCreated: Date) {
     /*3 collison 5 posision*/
 }
